Refactored Navigation component
