+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Navigation from "./components/Navigation";
+import Navigation from "./components/Navigation/Navigation";
 
 export default function App() {
   return <Navigation></Navigation>
